Add Hamiltonian generation method, remove file
The Hamiltonian method ensures that the pairs will be a cycle over the
entire group. This allows you to pick one person, and then continue to
pick the person they are paired with until everyone has been selected.

The Hamiltonian method will fail if there is no Hamiltonian for the
graph (duh!). A graph-based method to support topologies with no
Hamiltonian would be cool, but I don't need it and it sounds a little
complex.